mds/tools: fix mds->mdsmap->get_metadata_pool() return int->int64_t
mds->mdsmap->get_metadata_pool() return should be int64_t not int,
fix up

+++ src/tools/cephfs/JournalTool.cc
@@ -116,7 +116,7 @@ int JournalTool::main(std::vector<const char*> &argv)
  
   auto fs = fsmap->get_filesystem(role_selector.get_ns());
   assert(fs != nullptr);
-  int const pool_id = fs->mds_map.get_metadata_pool();
+  int64_t const pool_id = fs->mds_map.get_metadata_pool();
   dout(4) << "JournalTool: resolving pool " << pool_id << dendl;
   std::string pool_name;
   r = rados.pool_reverse_lookup(pool_id, &pool_name);
